Move API controllers to a new Namespace: ICDataManager.Controllers.API and delete default scaffolded weather controller
