+++ ICDataManager/Controllers/API/IngredientTypesController.cs
@@ -8,7 +8,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace ICDataManager.Controllers
+namespace ICDataManager.Controllers.API
 {
     [Route("api/[controller]")]
     [ApiController]
