+++ ICDataManager/Controllers/API/IngredientsController.cs
@@ -1,4 +1,5 @@
 ﻿using ICDataManager.Library.Data;
+using ICDataManager.Library.DataAccess;
 using ICDataManager.Library.Helpers;
 using ICDataManager.Library.Models;
 using Microsoft.AspNetCore.Http;
@@ -8,36 +9,53 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace ICDataManager.Controllers
+namespace ICDataManager.Controllers.API
 {
     [Route("api/[controller]")]
     [ApiController]
-    public class IngredientsCheckerController : ControllerBase
+    public class IngredientsController : ControllerBase
     {
         private readonly IIngredientData _ingredientData;
-        private readonly IDisplayHelper _displayHelper;
         private readonly IIngredientTypeData _ingredientsTypeData;
+        private readonly IDisplayHelper _displayHelper;
         private readonly IIngredientNameData _ingredientNameData;
 
-        public IngredientsCheckerController(IIngredientData ingredientData, IDisplayHelper displayHelper, IIngredientTypeData ingredientsTypeData, IIngredientNameData ingredientNameData)
+        public IngredientsController(IIngredientData ingredientData,
+                                     IIngredientTypeData ingredientsTypeData,
+                                     IDisplayHelper displayHelper,
+                                     IIngredientNameData _ingredientNameData)
         {
             _ingredientData = ingredientData;
-            _displayHelper = displayHelper;
             _ingredientsTypeData = ingredientsTypeData;
-            _ingredientNameData = ingredientNameData;
+            _displayHelper = displayHelper;
+            this._ingredientNameData = _ingredientNameData;
         }
 
-        [HttpPost]
-        public async Task<CheckedIngredientsModel> PostIngredientsNameForChecking(string[] ingredientsNames)
+        [HttpGet]
+        public async Task<List<DisplayIngredientModel>> Get()
         {
-            if (ingredientsNames.GetLength(0) == 0)
+
+            var ingredientsList = await _ingredientData.GetAll();
+            var typesList = await _ingredientsTypeData.GetAll();
+
+            List<DisplayIngredientModel> detailedIngredientsList = await _displayHelper.GetIngrediensListForDisplay(ingredientsList, typesList);
+
+
+
+            return detailedIngredientsList;
+        }
+
+        [HttpPost("checker")]
+        public async Task<CheckedIngredientsModel> CheckByNames(IEnumerable<string> names)
+        {
+            if (names == null || !names.Any())
             {
-                throw new ArgumentNullException("ingredientName", "list of names to check is empty");
+                throw new ArgumentNullException("names", "list of names to check is empty");
             }
 
             var dbIngredientsList = new List<DBIngredientModel>();
             var notFoundNames = new List<string>();
-            foreach (var name in ingredientsNames)
+            foreach (var name in names)
             {
                 var ingredientName = await _ingredientNameData.GetByName(name);
 
@@ -55,7 +73,7 @@ namespace ICDataManager.Controllers
                 }
 
                 var dbIngredient = await _ingredientData.GetById(ingredientName.IngredientId);
-                
+
                 // TODO: there has to be nicer way to get back Ingredients with their searched name, not MainName
                 dbIngredient.MainNameId = ingredientName.Id;
 
@@ -74,7 +92,7 @@ namespace ICDataManager.Controllers
 
 
             return output;
-
         }
+
     }
 }
